IDLE Thread and Thread State introduced
The threads now have a state "RUN" or "WAIT" and each thread can put itself to sleep going to WAIT state. The OS will then schedule a new thread according to the state, if WAIT it will be skipped. Before scheduling each state is updated and if it's time to wake up the thread will be put to RUN. 
The OS_Wait function is introduced, it calls OS_Sched at the end because if the thread goes to sleep the context must be switched immediately.