Global stack, virtual scheduler
The scheduler is not virtual and can be reimplemented by the user, which can access to global variables through convenient getter/setter functions. 

The stack is now global and part of the OS class so the user only cares about defining a thread and then registering it in the OS. No need to reserve stack manually.