User driven approach
There is only one instance of the OS, created as global object from the user. 
The only three global variables used by the OS are the active threads list and the two pointers, necessary for the asm context switch. 
The user must provide the SysTick interrupt in which it calls OS_Advance which increases the time, that is kept internally by the OS.